PyCode/Intelliwrite/requirements.txt
agno
qdrant-client
openai
pydantic
python-dotenv
sqlalchemy
pg8000
Flask
flask-cors
duckduckgo-search
langfuse
pymongo